Added more file types, classes and project opening function...

+++ src/hkxclasses/behavior/generators/hkbblendergenerator.h
@@ -37,7 +37,7 @@ private:
     qreal blendParameter;
     qreal minCyclicBlendParameter;
     qreal maxCyclicBlendParameter;
-    qint16 indexOfSyncMasterChild;
+    int indexOfSyncMasterChild;
     QString flags;
     bool subtractLastChild;
     QList <HkxObjectExpSharedPtr> children;
